FA-42 fix layout homepage
